Refactor CLI commands for Celery executor and add provider info section
