+++ providers/celery/src/airflow/providers/celery/cli/definition.py
